+++ ex.py
@@ -1,0 +1,8 @@
+a = 1
+
+try:
+    ...
+except Exception as a:
+    ...
+
+print(a)