+++ wemake_python_styleguide/logic/variables.py
@@ -6,14 +6,19 @@ from typing import Union
 VarDefinition = Union[ast.AST, ast.expr]
 
 
-def is_valid_block_variable_definition(node: VarDefinition) -> bool:
-    """Is used to check either block variables are correctly defined."""
+def _is_valid_single(node: VarDefinition) -> bool:
     if isinstance(node, ast.Name):
         return True
-
-    if isinstance(node, ast.Tuple):
-        for var_definition in node.elts:
-            if not isinstance(var_definition, ast.Name):
-                return False
+    if isinstance(node, ast.Starred) and isinstance(node.value, ast.Name):
         return True
     return False
+
+
+def is_valid_block_variable_definition(node: VarDefinition) -> bool:
+    """Is used to check either block variables are correctly defined."""
+    if isinstance(node, ast.Tuple):
+        for var_definition in node.elts:
+            if not _is_valid_single(var_definition):
+                return False
+        return True
+    return _is_valid_single(node)